requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.2
discord==1.7.3
discord.py==1.7.3
idna==3.2
insta-scrape==2.1.2
instagramy==4.4
multidict==5.1.0
nest-asyncio==1.5.1
oauthlib==3.1.1
prompt-toolkit==3.0.19
PySocks==1.7.1
python-dotenv==0.18.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
tweepy==3.8.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
wit==6.0.0
yarl==1.6.3
emoji==1.4.2
replit==3.2.4
prsaw==0.4.0